financial-services/document-research-agent/requirements.txt
langgraph==1.0.8
langchain==1.2.10
langchain-core==1.2.13
langchain-anthropic==1.3.3
langchain-text-splitters==1.1.0
deepagents==0.4.1
pydantic>=2.0.0
python-dotenv>=1.0.0
